+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 -   Add support for adding resource packs from Curse
 -   Fix Curse mod file selector dependencies not showing well
 -   When installing an instance, give focus to name field and select all when clicked
+-   [#255] Add tool to download and use recommended version of Java to use